+++ example/with-paths/utils/src/to-ignore.ts
@@ -1,2 +1,2 @@
-// Should be ignored due to --ignorePaths option
+// Should be ignored due to --excludePathsFromReport option
 export const unused_fun = () => 0;