+++ src/usage.ts
@@ -1,6 +1,17 @@
 export const USAGE = `
-    usage: ts-unused-exports path/to/tsconfig.json [file1.ts file2.ts] [--allowUnusedTypes][--excludeDeclarationFiles][--exitWithCount][--ignorePaths=path1;path2][--searchNamespaces][--showLineNumber]
-  
+    usage: ts-unused-exports path/to/tsconfig.json [file1.ts file2.ts] [options]
+    
+    where options are any of:
+      --allowUnusedTypes
+      --excludeDeclarationFiles
+      --excludePathsFromReport=path1;path2
+      --exitWithCount
+      --ignoreFiles=<regex>
+      --ignoreProductionFiles
+      --ignoreTestFiles
+      --searchNamespaces
+      --showLineNumber
+
     Note: if no file is specified after tsconfig, the files will be read from the
     tsconfig's "files" key which must be present.